+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Changed the type of `Format`. Now you write `Format r (a -> r)` instead
   of `Format a`.
+* Add `formatToString` function.